+++ crates/holochain/src/core/sys_validate/tests.rs
@@ -261,7 +261,8 @@ async fn verify_action_signature_test() {
     let (record_valid, cascade) = record_with_cascade(&keystore, Action::CreateLink(action)).await;
 
     let wrong_signature = Signature([1_u8; 64]);
-    let action_invalid = SignedActionHashed::new(record_valid.action().clone(), wrong_signature);
+    let action_invalid =
+        SignedActionHashed::new_unchecked(record_valid.action().clone(), wrong_signature);
     let record_invalid = Record::new(action_invalid, None);
 
     sys_validate_record(&record_valid, &cascade).await.unwrap();
