+++ crates/holochain/src/lib.rs
@@ -42,7 +42,9 @@ pub mod prelude {
     pub use holo_hash;
     pub use holochain_p2p::AgentPubKeyExt;
     pub use holochain_p2p::*;
-    pub use holochain_types::inline_zome::*;
     pub use holochain_types::prelude::*;
     pub use kitsune_p2p::*;
+
+    #[cfg(feature = "test_utils")]
+    pub use holochain_types::inline_zome::*;
 }
