+++ crates/holochain_cascade/Cargo.toml
@@ -34,19 +34,21 @@ thiserror = "1.0"
 tracing = "0.1"
 tracing-futures = "0.2"
 
-async-trait = { version = "0.1", optional = true }
+async-trait = { version = "0.1" }
 mockall = { version = "0.11.3", optional = true }
 
 [dev-dependencies]
+holochain_cascade = { path = ".", features = ["test_utils"] }
+
 isotest = "0"
 matches = "0.1"
 pretty_assertions = "0.7.2"
 test-case = "2.1"
 
 [features]
-default = ["test_utils"]
+default = []
 
-test_utils = ["async-trait", "mockall", "holochain_types/test_utils"]
+test_utils = ["mockall", "holochain_types/test_utils", "holochain_p2p/test_utils"]
 
 sqlite-encrypted = [
     "holo_hash/sqlite-encrypted",
