+++ crates/holochain_p2p/src/lib.rs
@@ -24,7 +24,7 @@ pub use test::HolochainP2pDnaFixturator;
 
 pub use kitsune_p2p;
 
-#[mockall::automock]
+#[cfg_attr(feature = "test_utils", mockall::automock)]
 #[allow(clippy::too_many_arguments)]
 #[async_trait::async_trait]
 /// A wrapper around HolochainP2pSender that partially applies the dna_hash / agent_pub_key.
