+++ crates/holochain_sqlite/Cargo.toml
@@ -71,7 +71,7 @@ loom = { version = "0.6", features = ["futures", "checkpoint"] }
 holochain_sqlite = { path = ".", default-features = false, features = ["test_utils"] }
 
 [features]
-default = [ "test_utils", "sqlite", "kitsune_p2p_types", "holochain_util" ]
+default = [ "sqlite", "kitsune_p2p_types", "holochain_util" ]
 
 test_utils = [ ]
