+++ crates/holochain_state/Cargo.toml
@@ -50,6 +50,8 @@ nanoid = { version = "0.3", optional = true }
 contrafact = { version = "0.2.0-rc.1", optional = true }
 
 [dev-dependencies]
+holochain_state = { path = ".", features = ["test_utils"] }
+
 anyhow = "1.0.26"
 arbitrary = "1.0"
 fixt = { path = "../fixt" }
@@ -62,7 +64,7 @@ pretty_assertions = "0.6.1"
 tempfile = "3.3"
 
 [features]
-default = ["test_utils"]
+default = []
 
 test_utils = [
     "holochain_types/test_utils",
