+++ crates/holochain_types/Cargo.toml
@@ -69,7 +69,7 @@ isotest = { version = "0", optional = true }
 contrafact = { version = "0.2.0-rc.1", optional = true }
 
 [dev-dependencies]
-holochain_types = { path = ".", features = ["test_utils"]}
+holochain_types = { path = ".", features = ["test_utils", "fuzzing"]}
 
 arbitrary = "1.0"
 isotest = { version = "0" }
@@ -81,7 +81,7 @@ test-case = "2.1.0"
 tokio = { version = "1.11", features = [ "full" ] }
 
 [features]
-default = ["fixturators", "test_utils"]
+default = []
 
 fixturators = ["holochain_zome_types/fixturators"]
 
@@ -92,7 +92,9 @@ fuzzing = [
   "holochain_zome_types/fuzzing",
   "mr_bundle/fuzzing",
 ]
+
 test_utils = [
+  "fixturators",
   "fuzzing",
   "isotest",
   "holochain_zome_types/test_utils",
