+++ crates/holochain_types/src/dna/dna_file.rs
@@ -1,6 +1,5 @@
 use super::error::DnaError;
 use crate::prelude::*;
-use arbitrary::Arbitrary;
 use holo_hash::*;
 use holochain_zome_types::ZomeName;
 use std::collections::BTreeMap;
@@ -21,8 +20,8 @@ mod test;
     derive_more::AsRef,
     derive_more::From,
     derive_more::IntoIterator,
-    Arbitrary,
 )]
+#[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
 #[serde(from = "WasmMapSerialized", into = "WasmMapSerialized")]
 pub struct WasmMap(BTreeMap<holo_hash::WasmHash, wasm::DnaWasm>);
 
@@ -53,7 +52,8 @@ impl From<WasmMapSerialized> for WasmMap {
 ///       we should remove the Serialize impl on this type, and perhaps rename
 ///       to indicate that this is simply a validated, fully-formed DnaBundle
 ///       (i.e. all Wasms are bundled and immediately available, not remote.)
-#[derive(Serialize, Deserialize, Clone, PartialEq, Eq, SerializedBytes, Hash, Arbitrary)]
+#[derive(Serialize, Deserialize, Clone, PartialEq, Eq, SerializedBytes, Hash)]
+#[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
 pub struct DnaFile {
     /// The hashable portion that can be shared with hApp code.
     pub(super) dna: DnaDefHashed,
