+++ crates/kitsune_p2p/fetch/Cargo.toml
@@ -32,8 +32,8 @@ arbitrary = "1"
 rand = "0.8.5"
 
 [features]
+default = []
 test_utils = ["human-repr"]
-default = ["test_utils"]
 sqlite-encrypted = [
     "kitsune_p2p_types/sqlite-encrypted",
 ]
