+++ crates/kitsune_p2p/types/Cargo.toml
@@ -59,6 +59,7 @@ fuzzing = [
 test_utils = [
   "kitsune_p2p_bin_data/test_utils",
   "kitsune_p2p_dht_arc/test_utils",
+  "kitsune_p2p_dht/test_utils",
   "ghost_actor/test_utils",
   "mockall",
 ]
